Add caching of login items for increased security
Instead of descrypting the whole bitwarden collection everytime, the
collection is decrypted once and stored on disk in a cache file after
stripping it of sensitive data such as passwords or totp data. The cache
is used as long as it's been generated in the last 'security.cache'
days. When a cache has expired, a refresh is automatically done.

Add --cache console argument, and 'security.cache' config key to control
the expiry time of the cache